requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pytesseract>=0.3.8
pillow>=8.3.0
pdf2image>=1.16.0
openpyxl>=3.0.0